scale: Fix up toggling between modes
This was broken by fa466b58 because scale was relying on this behavior
to determine if a view was on the current workspace. Now, wayfire includes
the transform when building the list of workspace views which means we are
left to do a bit of logic to determine if the view is on the current
workspace. The heuristic used is whether or not the center point of the
view is contained in the workspace. This fixes switching between modes
using the bindings.